[REF] html_editor: move savePendingImages out of media plugin
This commit moves image saving functionalities out of the Media plugin
into a dedicated plugin.

This allows us to have the editor without the Media plugin, but still
able to save pasted or dropped images as attachments.

task-5043921
