+++ addons/html_editor/static/src/fields/html_field.js
@@ -164,7 +164,7 @@ export class HtmlField extends Component {
     }
 
     async getEditorContent() {
-        await this.editor.shared.media?.savePendingImages();
+        await this.editor.shared.imageSave?.savePendingImages();
         return this.editor.getElContent();
     }
