+++ addons/project/static/src/project_sharing/editor/project_sharing_media_plugin.js
@@ -1,4 +1,5 @@
 import { ImageCropPlugin } from "@html_editor/main/media/image_crop_plugin";
+import { ImageSavePlugin } from "@html_editor/main/media/image_save_plugin";
 import { MediaPlugin } from "@html_editor/main/media/media_plugin";
 import { MAIN_PLUGINS } from "@html_editor/plugin_sets";
 
@@ -9,6 +10,9 @@ export class ProjectSharingMediaPlugin extends MediaPlugin {
             item => item.id !== "replace_image"
         ),
     }
+}
+
+export class ProjectSharingImageSavePlugin extends ImageSavePlugin {
     async createAttachment({ el, imageData, resId }) {
         const response = JSON.parse(
             await this.services.http.post(
@@ -35,5 +39,7 @@ export class ProjectSharingMediaPlugin extends MediaPlugin {
 
 MAIN_PLUGINS.splice(MAIN_PLUGINS.indexOf(MediaPlugin), 1);
 MAIN_PLUGINS.push(ProjectSharingMediaPlugin);
+MAIN_PLUGINS.splice(MAIN_PLUGINS.indexOf(ImageSavePlugin), 1);
+MAIN_PLUGINS.push(ProjectSharingImageSavePlugin);
 
 MAIN_PLUGINS.splice(MAIN_PLUGINS.indexOf(ImageCropPlugin), 1);
